🚸 Move link status to user menu
連携中のインストラクター名と解除操作をアバターメニューへ移しました。\nどの画面からでも連携状態を確認・解除できるようにしました。
