+++ src/routes/index.tsx
@@ -3,7 +3,6 @@ import { createFileRoute, redirect } from '@tanstack/react-router';
 
 import { ensureAuthenticated } from '@/features/auth/auth-guard';
 import { useMe } from '@/features/auth/queries';
-import { InstructorLinkStatus } from '@/features/dashboard/components/InstructorLinkStatus';
 import { UpcomingShifts } from '@/features/dashboard/components/UpcomingShifts';
 
 /**
@@ -27,12 +26,7 @@ function DashboardPage() {
     <Container size="sm" py="md">
       <Stack gap="md">
         <Title order={2}>ダッシュボード</Title>
-        {instructorId && (
-          <>
-            <UpcomingShifts instructorId={instructorId} />
-            <InstructorLinkStatus />
-          </>
-        )}
+        {instructorId && <UpcomingShifts instructorId={instructorId} />}
       </Stack>
     </Container>
   );
